+++ delphin/mrs/eds.py
@@ -413,14 +413,14 @@ def loads(s, single=False):
     return es
 
 
-def dump(fh, ms, single=False,
+def dump(destination, ms, single=False,
          properties=False, pretty_print=True,
          show_status=False, predicate_modifiers=False, **kwargs):
     """
     Serialize Xmrs objects to Eds and write to a file
 
     Args:
-        fh: file object where data will be written
+        destination: filename or file object where data will be written
         ms: an iterator of :class:`~delphin.mrs.xmrs.Xmrs` objects to
             serialize (unless the *single* option is `True`)
         single (bool): if `True`, treat *ms* as a single
@@ -431,14 +431,19 @@ def dump(fh, ms, single=False,
         show_status (bool): if `True`, annotate disconnected graphs and
             nodes
     """
-    print(dumps(ms,
-                single=single,
-                properties=properties,
-                pretty_print=pretty_print,
-                show_status=show_status,
-                predicate_modifiers=predicate_modifiers,
-                **kwargs),
-          file=fh)
+    text = dumps(ms,
+                 single=single,
+                 properties=properties,
+                 pretty_print=pretty_print,
+                 show_status=show_status,
+                 predicate_modifiers=predicate_modifiers,
+                 **kwargs)
+
+    if hasattr(destination, 'write'):
+        print(text, file=destination)
+    else:
+        with open(destination, 'w') as fh:
+            print(text, file=fh)
 
 
 def dumps(ms, single=False,