+++ delphin/mrs/simplemrs.py
@@ -99,13 +99,13 @@ def loads(s, single=False, version=_default_version,
         return ms
 
 
-def dump(fh, ms, single=False, version=_default_version, properties=True,
+def dump(destination, ms, single=False, version=_default_version, properties=True,
          pretty_print=False, color=False, **kwargs):
     """
     Serialize Xmrs objects to SimpleMRS and write to a file
 
     Args:
-        fh: file object where data will be written
+        destination: filename or file object where data will be written
         ms: an iterator of Xmrs objects to serialize (unless the
             *single* option is `True`)
         single: if `True`, treat *ms* as a single Xmrs object
@@ -114,14 +114,19 @@ def dump(fh, ms, single=False, version=_default_version, properties=True,
         pretty_print: if `True`, add newlines and indentation
         color: if `True`, colorize the output with ANSI color codes
     """
-    print(dumps(ms,
-                single=single,
-                version=version,
-                properties=properties,
-                pretty_print=pretty_print,
-                color=color,
-                **kwargs),
-          file=fh)
+    text = dumps(ms,
+                 single=single,
+                 version=version,
+                 properties=properties,
+                 pretty_print=pretty_print,
+                 color=color,
+                 **kwargs)
+
+    if hasattr(destination, 'write'):
+        print(text, file=destination)
+    else:
+        with open(destination, 'w') as fh:
+            print(text, file=fh)
 
 
 def dumps(ms, single=False, version=_default_version, properties=True,