+++ tests/mrs_compare_test.py
@@ -123,6 +123,24 @@ pathological2 = simplemrs.loads_one('''
   HCONS: < h0 qeq h1 h5 qeq h12 h9 qeq h11 h15 qeq h17 h22 qeq h29 h26 qeq h28 h32 qeq h34 h39 qeq h46 h43 qeq h45 h49 qeq h51 > ]
 ''')
 
+# x1 and x2 differ only by the presence of '_rel' at the end of the predicates
+
+x1 = simplemrs.loads_one('''
+[ LTOP: h0 INDEX: e2 [ e SF: prop TENSE: past MOOD: indicative PROG: - PERF: - ]
+  RELS: < [ proper_q<0:6> LBL: h4 ARG0: x3 [ x PERS: 3 NUM: sg IND: + ] RSTR: h5 BODY: h6 ]
+          [ named<0:6> LBL: h7 CARG: "Abrams" ARG0: x3 ]
+          [ _sleep_v_1<7:13> LBL: h1 ARG0: e2 ARG1: x3 ] >
+  HCONS: < h0 qeq h1 h5 qeq h7 > ]
+''')
+
+x2 = simplemrs.loads_one('''
+[ LTOP: h0 INDEX: e2 [ e SF: prop TENSE: past MOOD: indicative PROG: - PERF: - ]
+  RELS: < [ proper_q_rel<0:6> LBL: h4 ARG0: x3 [ x PERS: 3 NUM: sg IND: + ] RSTR: h5 BODY: h6 ]
+          [ named_rel<0:6> LBL: h7 CARG: "Abrams" ARG0: x3 ]
+          [ _sleep_v_1_rel<7:13> LBL: h1 ARG0: e2 ARG1: x3 ] >
+  HCONS: < h0 qeq h1 h5 qeq h7 > ]
+''')
+
 def test_isomorphic():
     assert compare.isomorphic(m1, m1) == True  # identity
     assert compare.isomorphic(m1, m1b) == True  # diff Lnk only
@@ -135,3 +153,8 @@ def test_isomorphic():
     # be aware if the next ones take a long time to resolve
     assert compare.isomorphic(pathological1, pathological1) == True
     assert compare.isomorphic(pathological1, pathological2) == False
+    # test for normalized forms being treated as equivalent by compare.isomorphic
+    assert x1 == x2 # generally a stricter test than isomorphism
+    assert compare.isomorphic(x1, x2) == True # if normalized
+    assert compare.isomorphic(x1, x1) == True # identity
+    assert compare.isomorphic(x2, x2) == True # identity